Add initial chains state service wrapper
